Minor clean up, add `utils.py` for general utilities - consider moving/renaming at some later point.
