+++ nengo_spinnaker/node_builders/ethernet.py
@@ -9,6 +9,7 @@ from pacman103.core.spinnman.sdp import sdp_message as sdp
 
 from . import receive_vertex, transmit_vertex
 from . import io_builder
+from .. import utils
 
 
 class Ethernet(io_builder.IOBuilder):
